Rearrange functions and remove necessary complexity
